requirements.txt
getmac
pyserial
mysql-connector
configparser
pygame
requests
PyMySQL
pyscard
PyQT5